rpi/requirements.txt
# Intelligent Pesticide Sprinkling System — RPi Python Dependencies
# Install: pip install -r requirements.txt --break-system-packages
#
# NOTE: picamera2 must be installed via apt, NOT pip:
#   sudo apt install python3-picamera2

# Gemini Vision API
google-generativeai>=0.5.0

# DHT22 sensor (1-Wire on GPIO4)
adafruit-circuitpython-dht>=3.7.0
RPi.GPIO>=0.7.1
board

# NPK sensor (Modbus RTU via MAX485 on GPIO14/15, DE/RE on GPIO17)
minimalmodbus>=2.1.0
pyserial>=3.5

# Image handling for Gemini
Pillow>=10.0.0

# Environment variable management
python-dotenv>=1.0.0
